backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["nurliman <nurlimandiara@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.1"
flask-jwt-extended = "^4.6.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
dataclasses-json = "^0.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
